itb-auth/src/main/java/com/chunkymonkey/itb/converter/GrantedAuthorityDeserializer.java
package com.chunkymonkey.itb.converter;

import java.io.IOException;

import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.authority.SimpleGrantedAuthority;

import com.fasterxml.jackson.core.JsonParser;
import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.DeserializationContext;
import com.fasterxml.jackson.databind.JsonDeserializer;

public class GrantedAuthorityDeserializer extends JsonDeserializer<GrantedAuthority> {

	@Override
	public GrantedAuthority deserialize(JsonParser p, DeserializationContext ctxt)
			throws IOException, JsonProcessingException {
		String value = p.getValueAsString();
		if (value == null || value == "")
			return null;
		return new SimpleGrantedAuthority(value);
	}

}
